requirements.txt
beautifulsoup4==4.12.2
requests==2.31.0
python-telegram-bot==20.4
python-dotenv~=1.0.0